+++ src/main/java/com/notenoughmail/kubejs_tfc/block/LogBlockBuilder.java
@@ -14,6 +14,7 @@ import org.jetbrains.annotations.Nullable;
 
 import java.util.function.Consumer;
 
+// TODO: Fix item model being empty under conditions of "log_test_with_stripped"
 public class LogBlockBuilder extends ExtendedPropertiesShapedBlockBuilder {
 
     @Nullable